services/categories_service/requirements.txt
Django==5.1.2
djangorestframework==3.16.0
dj-database-url==2.2.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
gunicorn==21.2.0
gspread==6.1.4

